feat: slash notes guild
