+++ src/services/notes-service.ts
@@ -1,5 +1,5 @@
 import Axios from 'axios';
-import { Message } from 'discord.js';
+import { CommandInteraction, Message, MessageEmbed } from 'discord.js';
 import { Notes } from '../models/notes';
 import { Command, Response } from '../models/command';
 import { GuildConfig } from '../models/guild';
@@ -8,12 +8,10 @@ import { Tokenizer } from '../util/tokenizer';
 import { Colors, EmbedBuilder } from '../util/embed-builder';
 
 export class NotesService {
-  private prefix: string;
-  private command: Command;
+  private interaction: CommandInteraction;
 
-  constructor(command: Command, prefix: string) {
-    this.command = command;
-    this.prefix = prefix;
+  constructor(interaction: CommandInteraction) {
+    this.interaction = interaction;
   }
 
   static async get(id: string): Promise<Notes | undefined> {
@@ -43,84 +41,123 @@ export class NotesService {
     }).then((res) => res.data);
   }
 
-  async response(msg: Message, guildConfig: GuildConfig | undefined): Promise<Response> {
-    const tokenizer = new Tokenizer(msg.content, this.prefix);
-
-    //!notes - get notes in channel
-    if (tokenizer.tokens.length === 1) {
-      if (guildConfig) {
-        return this.getByChannel(msg.channel.id.toString(), guildConfig.id);
+  // TODO: Fix for global/dms, maybe not and leave guild
+  async response(/*msg: Message, guildConfig: GuildConfig | undefined*/): Promise<void> {
+    if (this.interaction.guildId == null) {
+      return;
+    }
+    const channelInteraction = this.interaction.options.getChannel('channel', false);
+    switch(this.interaction.options.getSubcommand()){
+      case 'view': {
+        if (channelInteraction != null && channelInteraction.type === 'GUILD_TEXT') {
+          this.interaction.reply({ embeds: [ await this.getByChannel(channelInteraction.id, this.interaction.guildId)] });
+        }
+        else {
+          this.interaction.reply({ embeds: [ await this.getByChannel(this.interaction.channelId, this.interaction.guildId)] });
+        }
+        break;
       }
-      else {
-        return this.getByUser(msg.author.id);
+      case 'add': {
+        if (channelInteraction != null && channelInteraction.type === 'GUILD_TEXT') {
+          this.interaction.reply({ embeds: [ await this.setChannelNote(this.interaction.options.getString('note', true), channelInteraction.id, this.interaction.guildId)] });
+        }
+        else {
+          this.interaction.reply({ embeds: [ await this.setChannelNote(this.interaction.options.getString('note', true), this.interaction.channelId, this.interaction.guildId)] });
+        }
+        break;
+      }
+        break;
+      case 'remove': {
+        if (channelInteraction != null && channelInteraction.type === 'GUILD_TEXT') {
+          this.interaction.reply({ embeds: [ await this.delChannelNote(this.interaction.options.getInteger('note', true), channelInteraction.id, this.interaction.guildId)] });
+        }
+        else {
+          this.interaction.reply({ embeds: [ await this.delChannelNote(this.interaction.options.getInteger('note', true), this.interaction.channelId, this.interaction.guildId)] });
+        }
+        break;
       }
     }
 
-    if (guildConfig) {
-      //!notes #channel - get notes for a channel
-      if (tokenizer.tokens[1]?.type === 'channel') {
-        return this.getByChannel(tokenizer.tokens[1].content.substr(2, 18), guildConfig.id);
-      }
-      //!notes add #channel - adds this note
-      if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1]?.content === 'add' && tokenizer.tokens[2]?.type === 'channel' && tokenizer.tokens[3]?.type === 'text') {
-        return this.setChannelNote(tokenizer.body(3), tokenizer.tokens[2].content.substr(2, 18), guildConfig.id);
-      }
-      //!notes add
-      if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1]?.content === 'add' && tokenizer.tokens[2]?.type === 'text') {
-        return this.setChannelNote(tokenizer.body(2), msg.channel.id, guildConfig.id);
-      }
-    }
-    else {
-      // DM/user
-      if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1]?.content === 'add' && tokenizer.tokens[2]?.type === 'text') {
-        return this.setUserNote(tokenizer.body(2), msg.author.id);
-      }
-    }
+    // ////////
+    // const tokenizer = new Tokenizer(msg.content, this.prefix);
 
-    //!notes remove <channel> <number>
-    if (guildConfig) {
-      if (!(msg.member?.permissions.has(['ADMINISTRATOR'], true))) {
-        return 'You have to be an admin to delete notes.';
-      }
+    // //!notes - get notes in channel
+    // if (tokenizer.tokens.length === 1) {
+    //   if (guildConfig) {
+    //     return this.getByChannel(msg.channel.id.toString(), guildConfig.id);
+    //   }
+    //   else {
+    //     return this.getByUser(msg.author.id);
+    //   }
+    // }
 
-      if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1].content === 'remove'
-        && tokenizer.tokens[2]?.type === 'channel' && tokenizer.tokens[3]?.type === 'text') {
-        const noteNum: number = parseInt(tokenizer.tokens[3].content);
-        return this.delChannelNote(noteNum, tokenizer.tokens[2].content.substr(2, 18), guildConfig.id);
-      }
+    // if (guildConfig) {
+    //   //!notes #channel - get notes for a channel
+    //   if (tokenizer.tokens[1]?.type === 'channel') {
+    //     return this.getByChannel(tokenizer.tokens[1].content.substr(2, 18), guildConfig.id);
+    //   }
+    //   //!notes add #channel - adds this note
+    //   if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1]?.content === 'add' && tokenizer.tokens[2]?.type === 'channel' && tokenizer.tokens[3]?.type === 'text') {
+    //     return this.setChannelNote(tokenizer.body(3), tokenizer.tokens[2].content.substr(2, 18), guildConfig.id);
+    //   }
+    //   //!notes add
+    //   if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1]?.content === 'add' && tokenizer.tokens[2]?.type === 'text') {
+    //     return this.setChannelNote(tokenizer.body(2), msg.channel.id, guildConfig.id);
+    //   }
+    // }
+    // else {
+    //   // DM/user
+    //   if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1]?.content === 'add' && tokenizer.tokens[2]?.type === 'text') {
+    //     return this.setUserNote(tokenizer.body(2), msg.author.id);
+    //   }
+    // }
 
-      if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1].content === 'remove' &&
-        tokenizer.tokens[2]?.type === 'text') {
-        const noteNum: number = parseInt(tokenizer.tokens[2].content);
-        return this.delChannelNote(noteNum, msg.channel.id, guildConfig.id);
-      }
-    }
-    else {
-      if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1].content === 'remove' && tokenizer.tokens[2]?.type === 'text') {
-        const noteNum: number = parseInt(tokenizer.tokens[2].content);
-        return this.delUserNote(noteNum, msg.author.id);
-      }
-    }
+    // ////////
+
+    // //!notes remove <channel> <number>
+    // if (guildConfig) {
+    //   if (!(msg.member?.permissions.has(['ADMINISTRATOR'], true))) {
+    //     return 'You have to be an admin to delete notes.';
+    //   }
+
+    //   if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1].content === 'remove'
+    //     && tokenizer.tokens[2]?.type === 'channel' && tokenizer.tokens[3]?.type === 'text') {
+    //     const noteNum: number = parseInt(tokenizer.tokens[3].content);
+    //     return this.delChannelNote(noteNum, tokenizer.tokens[2].content.substr(2, 18), guildConfig.id);
+    //   }
+
+    //   if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1].content === 'remove' &&
+    //     tokenizer.tokens[2]?.type === 'text') {
+    //     const noteNum: number = parseInt(tokenizer.tokens[2].content);
+    //     return this.delChannelNote(noteNum, msg.channel.id, guildConfig.id);
+    //   }
+    // }
+    // else {
+    //   if (tokenizer.tokens[1]?.type === 'text' && tokenizer.tokens[1].content === 'remove' && tokenizer.tokens[2]?.type === 'text') {
+    //     const noteNum: number = parseInt(tokenizer.tokens[2].content);
+    //     return this.delUserNote(noteNum, msg.author.id);
+    //   }
+    // }
     // When incorrectly used (includes !notes help)
-    return EmbedBuilder.buildHelp(this.command, this.prefix, Colors.error, {
-      '#channel (optional)': 'get notes from a channel or DM.',
-      'add #channel (optional)': 'enter a note in a channel or DM.',
-      'remove #channel (optional) notenumber': 'remove a note in a channel or DM.'
-    }, ['add #cooking-stuff Eggs', 'add a note here', 'remove #cooking-stuff 1']);
+    // return EmbedBuilder.buildHelp(this.command, this.prefix, Colors.error, {
+    //   '#channel (optional)': 'get notes from a channel or DM.',
+    //   'add #channel (optional)': 'enter a note in a channel or DM.',
+    //   'remove #channel (optional) notenumber': 'remove a note in a channel or DM.'
+    // }, ['add #cooking-stuff Eggs', 'add a note here', 'remove #cooking-stuff 1']);
   }
 
 
-  async getByChannel(channelID: string, guildID: string): Promise<Response> {
+  async getByChannel(channelID: string, guildID: string): Promise<MessageEmbed> {
     const notes = await NotesService.get(guildID);
     if (notes == null || Array.isArray(notes.notes) || notes.notes[channelID]?.length == 0) {
-      return EmbedBuilder.info('No notes found for this channel.', 'For help: notes help', 'No notes');
+      return EmbedBuilder.info('No notes found for this channel.', 'Add some notes :D', 'No notes');
     }
     else {
       return EmbedBuilder.buildList(Colors.success, 'Notes :memo:', notes.notes[channelID], `Notes for channel <#${channelID}>`);
     }
   }
 
-  async setChannelNote(note: string, channelID: string, guildID: string): Promise<Response> {
+  async setChannelNote(note: string, channelID: string, guildID: string): Promise<MessageEmbed> {
     const notes = await NotesService.get(guildID);
 
     // When a guild doesnt have notes at all yet
@@ -158,7 +195,7 @@ export class NotesService {
       });
   }
 
-  async delChannelNote(noteNum: number, channelID: string, guildID: string): Promise<Response> {
+  async delChannelNote(noteNum: number, channelID: string, guildID: string): Promise<MessageEmbed> {
     const notes = await NotesService.get(guildID);
 
     if (notes == null) {
@@ -185,17 +222,17 @@ export class NotesService {
     }
   }
 
-  async getByUser(userID: string): Promise<Response> {
+  async getByUser(userID: string): Promise<MessageEmbed> {
     const notes = await NotesService.get(userID);
     if (notes == null || !Array.isArray(notes.notes) || notes.notes.length == 0) {
-      return EmbedBuilder.info('No notes personal notes found.', 'For help: notes help', 'No notes');
+      return EmbedBuilder.info('No notes personal notes found.', 'Add some notes :D', 'No notes');
     }
     else {
       return EmbedBuilder.buildList(Colors.success, 'Your personal notes! :memo:', notes.notes);
     }
   }
 
-  async setUserNote(note: string, userID: string): Promise<Response> {
+  async setUserNote(note: string, userID: string): Promise<MessageEmbed> {
     const notes = await NotesService.get(userID);
 
     // When a user doesnt have notes at all yet
@@ -226,7 +263,7 @@ export class NotesService {
       });
   }
 
-  async delUserNote(noteNum: number, userID: string): Promise<Response> {
+  async delUserNote(noteNum: number, userID: string): Promise<MessageEmbed> {
     const notes = await NotesService.get(userID);
 
     if (notes == null) {
